Fixes for Auth logic
