+++ apps/native/app/signup.tsx
@@ -3,8 +3,8 @@ import { StyleSheet, Text, View } from "react-native";
 import { Link, useRouter } from "expo-router";
 import { AuthForm, type AuthFormValues } from "@repo/ui/src";
 import { useUser } from "@repo/context/src";
-
-const API_URL = "https://web-and-native-expense-tracker.vercel.app/api";
+import { fetch } from "expo/fetch";
+import process from "process";
 
 export default function SignupScreen() {
   const router = useRouter();
@@ -12,7 +12,7 @@ export default function SignupScreen() {
   const [errorMessage, setErrorMessage] = React.useState("");
 
   const onSignup = async (values: AuthFormValues) => {
-    const response = await fetch(`${API_URL}/signup`, {
+    const response = await fetch(`${process.env.API_URL}/signup`, {
       method: "POST",
       body: JSON.stringify({
         email: values.email,
