+++ apps/web/app/api/signup/route.ts
@@ -1,6 +1,6 @@
 import { createClient } from "../../../utils/supabase/server";
 
-export async function POST(req) {
+export async function POST(req: Request) {
   try {
     const { email, password } = await req.json();
     const supabase = createClient();
@@ -21,7 +21,7 @@ export async function POST(req) {
       status: 200,
       headers: { "Content-Type": "application/json" },
     });
-  } catch (error) {
+  } catch (error: any) {
     return new Response(JSON.stringify({ error: error.message }), {
       status: 500,
       headers: { "Content-Type": "application/json" },
