+++ apps/web/components/SignupForm.tsx
@@ -1,8 +1,48 @@
 "use client";
 
 import Link from "next/link";
-import { SignupForm as Form } from "@repo/ui/src";
 import { useRouter } from "next/navigation";
+
+import React, { useState } from "react";
+import { AuthForm, type AuthFormValues } from "@repo/ui";
+import { useUser } from "@repo/context/src";
+import * as process from "process";
+
+export function Form({ onSubmit }: { onSubmit: () => void }) {
+  const { setUser } = useUser();
+  const [errorMessage, setErrorMessage] = useState("");
+
+  const onSignup = async (values: AuthFormValues) => {
+    const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/signup`, {
+      method: "POST",
+      body: JSON.stringify({
+        email: values.email,
+        password: values.password,
+      }),
+    });
+    const data = await response.json();
+
+    if (!response.ok) {
+      setErrorMessage(
+        data?.error ||
+          "Something went wrong. Please try again in a few minutes.",
+      );
+      return;
+    }
+
+    setUser({ email: values.email, name: values.email });
+    onSubmit && onSubmit();
+  };
+
+  return (
+    <AuthForm
+      includeConfirmPassword
+      submitButtonText="Sign Up"
+      onSubmit={onSignup}
+      errorMessage={errorMessage}
+    />
+  );
+}
 
 export function SignupForm() {
   const router = useRouter();
