+++ packages/ui/src/auth/index.tsx
@@ -1,3 +1,1 @@
 export { AuthForm, type AuthFormProps, type AuthFormValues } from "./AuthForm";
-export { LoginForm, type LoginFormProps } from "./LoginForm";
-export { SignupForm, type SignupFormProps } from "./SignupForm";
